🚨 Fix warning

+++ src/Save.cpp
@@ -13,12 +13,12 @@ void save_png(
     Size::DataType               width,
     Size::DataType               height,
     const void*                  data,
-    int                          channels_count,
+    size_t                       channels_count,
     bool                         flip_vertically
 )
 {
     stbi_flip_vertically_on_write(flip_vertically ? 1 : 0);
-    stbi_write_png(file_path.string().c_str(), static_cast<int>(width), static_cast<int>(height), channels_count, data, 0);
+    stbi_write_png(file_path.string().c_str(), static_cast<int>(width), static_cast<int>(height), static_cast<int>(channels_count), data, 0);
 }
 
 auto save_png_to_string(Image const& image, bool flip_vertically) -> std::string
@@ -36,14 +36,14 @@ auto save_png_to_string(
     Size::DataType width,
     Size::DataType height,
     const void*    data,
-    int            channels_count,
+    size_t         channels_count,
     bool           flip_vertically
 ) -> std::string
 {
     stbi_flip_vertically_on_write(flip_vertically ? 1 : 0);
 
     std::string res{};
-    stbi_write_png_to_func(&write_to_string, &res, static_cast<int>(width), static_cast<int>(height), channels_count, data, 0);
+    stbi_write_png_to_func(&write_to_string, &res, static_cast<int>(width), static_cast<int>(height), static_cast<int>(channels_count), data, 0);
     return res;
 }
 
@@ -57,12 +57,12 @@ void save_jpeg(
     Size::DataType               width,
     Size::DataType               height,
     void const*                  data,
-    int                          channels_count,
+    size_t                       channels_count,
     bool                         flip_vertically
 )
 {
     stbi_flip_vertically_on_write(flip_vertically ? 1 : 0);
-    stbi_write_jpg(file_path.string().c_str(), static_cast<int>(width), static_cast<int>(height), channels_count, data, 100);
+    stbi_write_jpg(file_path.string().c_str(), static_cast<int>(width), static_cast<int>(height), static_cast<int>(channels_count), data, 100);
 }
 
 } // namespace img
